Trim Service
A new service which can be run on SVM or a linux based UVM.
The service would run fstrim command on the disks periodically.
fstrim will be run in small chunks but continuously in the background.
For SVMs it will run on disk models which are marked as discard capable
in hcl.json. For UVMs it will run on all the disks backed by a nutanix vdisk.
This patch contains all the code necessary to run the fstrim service.